Fix wrong variable type

+++ src/redis-cli.c
@@ -452,7 +452,7 @@ static uint64_t dictSdsHash(const void *key) {
 static int dictSdsKeyCompare(void *privdata, const void *key1,
         const void *key2)
 {
-    int l1,l2;
+    size_t l1,l2;
     DICT_NOTUSED(privdata);
 
     l1 = sdslen((sds)key1);
